requirements.txt
asgiref==3.9.1
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
numpy==2.3.2
sqlparse==0.5.3
trimesh==4.7.1
whitenoise==6.9.0
